change medium int to int in ipacc

+++ sql_db.cpp
@@ -705,7 +705,7 @@ void SqlDb_mysql::createSchema() {
   `saddr` int(32) unsigned NOT NULL,\
   `daddr` int(32) unsigned NOT NULL,\
   `proto` smallint(4) unsigned NOT NULL,\
-  `octects` mediumint(32) unsigned NOT NULL,\
+  `octects` int(32) unsigned NOT NULL,\
   `numpackets` mediumint(32) unsigned NOT NULL,\
   `interval` varchar(255) NULL DEFAULT NULL,\
   KEY `saddr` (`saddr`),\
@@ -721,7 +721,7 @@ void SqlDb_mysql::createSchema() {
   `sport` smallint(4) unsigned NOT NULL,\
   `dport` smallint(4) unsigned NOT NULL,\
   `proto` smallint(4) unsigned NOT NULL,\
-  `octects` mediumint(32) unsigned NOT NULL,\
+  `octects` int(32) unsigned NOT NULL,\
   `numpackets` mediumint(32) unsigned NOT NULL,\
   `interval` varchar(255) NULL DEFAULT NULL,\
   KEY `saddr` (`saddr`),\
